otimização do código

+++ Blackboard/src/br/unifor/contributor/Contributor.java
@@ -2,8 +2,7 @@ package br.unifor.contributor;
 
 public interface Contributor {
 	
-	
-	public Double execCondition();
+	public Integer execCondition();
 	
 	public void execAction();
